refactor(mcp): extract read-only helper module
