+++ openspec/changes/refactor-mcp-read-only-helper-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-read-only-helper-module
+
+Extract MCP read-only helper into a dedicated module
