+++ openspec/changes/refactor-mcp-read-only-helper-module/tasks.md
@@ -1,0 +1,7 @@
+## 1. Implementation
+- [x] Extract MCP read-only helper into `src/mcp/tools/read_only.rs`
+- [x] Keep MCP tool behavior and JSON envelopes unchanged
+
+## 2. Verification
+- [x] `cargo fmt --all`
+- [x] `just check`
