+++ src/mcp/tools.rs
@@ -1,10 +1,7 @@
-use anyhow::Context as _;
 use rmcp::{
     ErrorData as McpError,
     model::{CallToolRequestParam, CallToolResult, Content, Tool},
 };
-
-use crate::user_error::UserError;
 
 use super::confirm::{CONFIRM_TOKEN_TTL, ConfirmTokenBinding};
 use super::{AgentpackMcp, confirm};
@@ -18,6 +15,7 @@ mod evolve_propose;
 mod evolve_restore;
 mod explain;
 mod preview;
+mod read_only;
 mod rollback;
 mod status;
 mod tool_schema;
@@ -151,58 +149,6 @@ pub(super) struct ExplainArgs {
     pub kind: ExplainKindArg,
 }
 
-async fn call_read_only_in_process(
-    command: &'static str,
-    args: CommonArgs,
-) -> anyhow::Result<(String, serde_json::Value)> {
-    tokio::task::spawn_blocking(move || {
-        let repo_override = args.repo.as_ref().map(std::path::PathBuf::from);
-        let profile = args.profile.as_deref().unwrap_or("default");
-        let target = args.target.as_deref().unwrap_or("all");
-        let machine_override = args.machine.as_deref();
-
-        let result = crate::handlers::read_only::read_only_context(
-            repo_override.as_deref(),
-            machine_override,
-            profile,
-            target,
-        );
-
-        let (text, envelope) = match result {
-            Ok(crate::handlers::read_only::ReadOnlyContext {
-                targets,
-                plan,
-                warnings,
-                ..
-            }) => {
-                let data = crate::app::plan_json::plan_json_data(profile, targets, plan);
-                let mut envelope = crate::output::JsonEnvelope::ok(command, data);
-                envelope.warnings = warnings;
-                let text = serde_json::to_string_pretty(&envelope)?;
-                let envelope = serde_json::to_value(&envelope)?;
-                (text, envelope)
-            }
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error(command, &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                (text, envelope)
-            }
-        };
-
-        Ok((text, envelope))
-    })
-    .await
-    .context("mcp read-only handler task join")?
-}
-
 async fn call_doctor_in_process(args: DoctorArgs) -> anyhow::Result<(String, serde_json::Value)> {
     doctor::call_doctor_in_process(args).await
 }
@@ -315,7 +261,7 @@ pub(super) async fn call_tool(
     match request.name.as_ref() {
         "plan" => {
             let args = deserialize_args::<CommonArgs>(request.arguments)?;
-            let (text, envelope) = match call_read_only_in_process("plan", args).await {
+            let (text, envelope) = match read_only::call_read_only_in_process("plan", args).await {
                 Ok(v) => v,
                 Err(err) => {
                     return Ok(CallToolResult::structured_error(envelope_error(
@@ -330,7 +276,7 @@ pub(super) async fn call_tool(
         }
         "diff" => {
             let args = deserialize_args::<CommonArgs>(request.arguments)?;
-            let (text, envelope) = match call_read_only_in_process("diff", args).await {
+            let (text, envelope) = match read_only::call_read_only_in_process("diff", args).await {
                 Ok(v) => v,
                 Err(err) => {
                     return Ok(CallToolResult::structured_error(envelope_error(
